# openapi.yaml
openapi: 3.1.0
info:
  title: MUSINSA Personal Shopper Plugin
  version: 0.4.0
  description: Conversational shopping API over MUSINSA product ontology data with intent parsing, recommendations, comparison, and session shortlists.
servers:
  - url: http://localhost:8787
paths:
  /health:
    get:
      operationId: health
      summary: Health check
      responses:
        '200':
          description: OK
  /products/search:
    post:
      operationId: searchProducts
      summary: Search MUSINSA products by natural language and structured filters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductSearchRequest'
      responses:
        '200':
          description: Product search results with retrieval metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSearchResponse'
  /products/{productId}:
    get:
      operationId: getProductDetail
      summary: Get product detail from ontology store
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Product detail
        '404':
          description: Product not found
  /products/{productId}/reviews/summary:
    post:
      operationId: summarizeProductReviews
      summary: Summarize purchase-relevant fit/material/review signals
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                focus:
                  type: array
                  items: { type: string }
                customer_profile:
                  $ref: '#/components/schemas/CustomerProfile'
      responses:
        '200':
          description: Review and fit summary
  /shopper/recommend:
    post:
      operationId: recommendByProfile
      summary: Recommend products based on parsed shopping intent, profile, budget, and style preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendRequest'
      responses:
        '200':
          description: Recommendations with parsed_intent, assistant_summary, shortlist, and next questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendResponse'
  /shopper/compare:
    post:
      operationId: compareProducts
      summary: Compare multiple products for purchase decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_ids]
              properties:
                product_ids:
                  type: array
                  items: { type: string }
                criteria:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Product comparison table, best pick, and decision notes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResponse'
  /shopper/shortlist:
    post:
      operationId: saveShortlist
      summary: Save product candidates into an in-memory session shortlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveShortlistRequest'
      responses:
        '200':
          description: Saved shortlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shortlist'
  /shopper/shortlist/{sessionId}:
    get:
      operationId: getShortlist
      summary: Get a session shortlist
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Session shortlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shortlist'
    delete:
      operationId: clearShortlist
      summary: Clear a session shortlist
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Clear result
  /analytics/events:
    post:
      operationId: recordAnalyticsEvent
      summary: Record sanitized non-PII shopper behavior event for ontology/marketing analytics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEventRequest'
      responses:
        '200':
          description: Sanitized event accepted
        '400':
          description: Invalid JSON or event payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Analytics consent required when ANALYTICS_CONSENT_REQUIRED is enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      summary: Return aggregate non-PII usage, click-through, conversion, query, and intent statistics
      responses:
        '200':
          description: Analytics summary
  /analytics/notice:
    get:
      operationId: getAnalyticsNotice
      summary: Return consent/notice text for non-PII analytics collection
      responses:
        '200':
          description: Analytics notice
  /analytics/funnel:
    get:
      operationId: getAnalyticsFunnel
      summary: Return funnel, CTR, CVR, and event count metrics
      responses:
        '200':
          description: Funnel analytics
  /analytics/products:
    get:
      operationId: getAnalyticsProducts
      summary: Return top products, clicked products, and converted products
      responses:
        '200':
          description: Product analytics
  /analytics/queries:
    get:
      operationId: getAnalyticsQueries
      summary: Return top sanitized search/query patterns
      responses:
        '200':
          description: Query analytics
  /analytics/intents:
    get:
      operationId: getAnalyticsIntents
      summary: Return intent statistics for colors, categories, genders, and budgets
      responses:
        '200':
          description: Intent analytics
  /analytics/insights:
    get:
      operationId: getAnalyticsInsights
      summary: Return generated marketing and ontology-gap insights from non-PII shopper analytics
      responses:
        '200':
          description: Analytics insights
  /analytics/export:
    post:
      operationId: exportAnalyticsSummary
      summary: Regenerate local analytics summary for ontology export automation
      responses:
        '200':
          description: Export-ready analytics summary
components:
  schemas:
    ProductSearchRequest:
      type: object
      properties:
        query: { type: string }
        category: { type: string }
        brand: { type: string }
        price_min: { type: number }
        price_max: { type: number }
        gender: { type: string }
        style_tags:
          type: array
          items: { type: string }
        limit: { type: integer, default: 10 }
        retrieval_mode:
          type: string
          enum: [local_index, hybrid, opencrab_first]
          default: local_index
          description: local_index uses the precomputed in-memory catalog index; hybrid reranks OpenCrab/ontology candidate ids when provided.
        candidate_product_ids:
          type: array
          description: Candidate IDs from an external ontology retrieval step to rerank locally.
          items: { type: string }
        opencrab_candidate_product_ids:
          type: array
          description: Alias for candidate_product_ids when candidates came from OpenCrab ontology-pack retrieval.
          items: { type: string }
        opencrab_top_k:
          type: integer
          default: 80
          description: Optional top-k hint for configured OpenCrab retrieval adapter.
        candidate_limit:
          type: integer
          description: Alias top-k hint for ontology candidate retrieval adapters.
    ProductSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
        retrieval:
          type: object
          properties:
            mode: { type: string }
            candidate_source:
              type: array
              items: { type: string }
            local_index_used: { type: boolean }
            opencrab_candidates_used: { type: boolean }
            candidate_count: { type: integer }
            matched_count: { type: integer }
            cache_hit: { type: boolean }
            lexicon:
              type: object
            latency_ms:
              type: object
            opencrab_adapter:
              type: object
              description: Adapter status when OPENCRAB_RETRIEVAL_URL or OPENCRAB_RETRIEVAL_CACHE_PATH is configured.
              properties:
                source: { type: string }
                skipped: { type: boolean }
                cache_hit: { type: boolean }
                product_id_count: { type: integer }
                raw_count: { type: integer }
                latency_ms: { type: number }
                error: { type: string }
    CustomerProfile:
      type: object
      properties:
        height: { type: number }
        weight: { type: number }
        usual_size: { type: string }
        fit_preference: { type: string }
        style_preference:
          type: array
          items: { type: string }
    RecommendRequest:
      type: object
      properties:
        query: { type: string }
        occasion: { type: string }
        budget: { type: number }
        category:
          type: string
          description: Optional exact category path constraint, e.g. '상의 > 후드 티셔츠'.
        brand: { type: string }
        customer_profile:
          $ref: '#/components/schemas/CustomerProfile'
        limit: { type: integer, default: 5 }
    ParsedIntent:
      type: object
      properties:
        raw_query: { type: string }
        budget: { type: number }
        colors:
          type: array
          items: { type: string }
        categories:
          type: array
          items: { type: string }
        seasons:
          type: array
          items: { type: string }
        gender: { type: string }
        brand: { type: string }
        must_have_terms:
          type: array
          items: { type: string }
    ProductCard:
      type: object
      properties:
        product_id: { type: string }
        product_name: { type: string }
        price: { type: number }
        score_breakdown:
          $ref: '#/components/schemas/ScoreBreakdown'
        reason: { type: string }
        url: { type: string }
        image: { type: string }
    ScoreBreakdown:
      type: object
      description: Transparent recommendation score components. Values are explanatory signals, not a regulated credit/risk score.
      properties:
        total: { type: number }
        search_score: { type: number }
        intent_match: { type: number }
        price_fit: { type: number }
        review_trust: { type: number }
        style_context: { type: number }
        personalization: { type: number }
        business_signal: { type: number }
        risk_penalty: { type: number }
        matched_signals:
          type: object
        explanation:
          type: object
    RecommendResponse:
      type: object
      properties:
        query_understood: { type: string }
        parsed_intent:
          $ref: '#/components/schemas/ParsedIntent'
        recommendation_confidence:
          type: object
          properties:
            level: { type: string }
            top_score: { type: number }
            low_confidence: { type: boolean }
            missing_ontology_fields:
              type: array
              items: { type: string }
            recommendation_count: { type: integer }
        assistant_summary: { type: string }
        recommendations:
          type: array
          items: { type: object }
        shortlist:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
        next_questions:
          type: array
          items: { type: string }
    CompareResponse:
      type: object
      properties:
        comparison_table:
          type: array
          items: { type: object }
        best_pick:
          type: object
        decision_notes:
          type: array
          items: { type: string }
    SaveShortlistRequest:
      type: object
      properties:
        session_id: { type: string }
        product_ids:
          type: array
          items: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
    Shortlist:
      type: object
      properties:
        session_id: { type: string }
        item_count: { type: integer }
        created_at: { type: string }
        updated_at: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
    AnalyticsEventRequest:
      type: object
      description: Sanitized non-PII event. Do not send names, emails, phone numbers, addresses, raw order IDs, IP addresses, or raw user identifiers.
      properties:
        event_type:
          type: string
          enum: [search, recommendation, product_click, shortlist_save, compare, purchase_intent, conversion, low_confidence_recommendation]
        session_id:
          type: string
          description: Anonymous session identifier; server stores only a SHA-256 hash prefix.
        query: { type: string }
        parsed_intent:
          $ref: '#/components/schemas/ParsedIntent'
        product_ids:
          type: array
          items: { type: string }
        clicked_product_id: { type: string }
        converted_product_id: { type: string }
        rank: { type: integer }
        confidence: { type: number }
        missing_ontology_fields:
          type: array
          items: { type: string }
        source: { type: string }
        consent_granted:
          type: boolean
          description: Set true after the user has accepted the analytics notice. Required when ANALYTICS_CONSENT_REQUIRED=true.
        consent_notice_version:
          type: string
          description: Analytics notice version shown to the user.
        consent:
          type: object
          properties:
            granted: { type: boolean }
            notice_version: { type: string }
            mode: { type: string }
        metadata:
          type: object
          properties:
            surface: { type: string }
            locale: { type: string }
            experiment: { type: string }
            result_count: { type: number }
            latency_ms: { type: number }
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }